backend-integration-tests/pyproject.toml
[project]
name = "backend-integration-tests"
version = "0.1.0"
description = "Integration tests for carpaintr backend service"
requires-python = ">=3.11"
dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["tests"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--capture=no",
]
markers = [
    "auth: Authentication tests",
    "admin: Admin-specific tests",
    "license: License management tests",
    "integration: Integration tests requiring running backend",
]

[tool.coverage.run]
source = ["tests"]
omit = ["tests/conftest.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]
